[sequentialVertexColoring] Renamed directory from graphColoring to coloring
