+++ configuration.conf
@@ -41,7 +41,7 @@ topologicalSort     | Y | Y | Y
 transitiveClosure   | Y | Y | Y
 breadthFirstSearch  | Y | Y | Y
 depthFirstSearch    | Y | Y | Y
-graphColoring       | Y | Y | Y
+coloring            | Y | Y | Y
 #----------------------
 # SQL only directories
 #----------------------
