+++ doc/coloring/CMakeLists.txt
