+++ doc/coloring/pgr_sequentialVertexColoring.rst
