+++ docqueries/coloring/CMakeLists.txt
