+++ docqueries/coloring/doc-pgr_sequentialVertexColoring.result
