+++ docqueries/coloring/doc-pgr_sequentialVertexColoring.test.sql
