+++ docqueries/coloring/test.conf
