+++ include/drivers/coloring/sequentialVertexColoring_driver.h
@@ -27,8 +27,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
  ********************************************************************PGR-GNU*/
 
-#ifndef INCLUDE_DRIVERS_GRAPHCOLORING_SEQUENTIALVERTEXCOLORING_DRIVER_H_
-#define INCLUDE_DRIVERS_GRAPHCOLORING_SEQUENTIALVERTEXCOLORING_DRIVER_H_
+#ifndef INCLUDE_DRIVERS_COLORING_SEQUENTIALVERTEXCOLORING_DRIVER_H_
+#define INCLUDE_DRIVERS_COLORING_SEQUENTIALVERTEXCOLORING_DRIVER_H_
 #pragma once
 
 /* for size-t */
@@ -67,4 +67,4 @@ extern "C" {
 }
 #endif
 
-#endif  // INCLUDE_DRIVERS_GRAPHCOLORING_SEQUENTIALVERTEXCOLORING_DRIVER_H_
+#endif  // INCLUDE_DRIVERS_COLORING_SEQUENTIALVERTEXCOLORING_DRIVER_H_
