+++ pgtap/coloring/sequentialVertexColoring/edge_cases.sql
