+++ pgtap/coloring/sequentialVertexColoring/inner_query.sql
