+++ pgtap/coloring/sequentialVertexColoring/no_crash_test.sql
