+++ pgtap/coloring/sequentialVertexColoring/types_check.sql
