+++ sql/coloring/CMakeLists.txt
