+++ sql/coloring/_sequentialVertexColoring.sql
