+++ sql/coloring/sequentialVertexColoring.sql
