+++ src/coloring/sequentialVertexColoring.c
@@ -42,7 +42,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_common/edges_input.h"
 #include "c_common/arrays_input.h"
 
-#include "drivers/graphColoring/sequentialVertexColoring_driver.h"
+#include "drivers/coloring/sequentialVertexColoring_driver.h"
 
 PGDLLEXPORT Datum _pgr_sequentialvertexcoloring(PG_FUNCTION_ARGS);
 PG_FUNCTION_INFO_V1(_pgr_sequentialvertexcoloring);
